added lithuanian language; translations moved into i18n dir
