+++ lib/i18n/fr.js
