+++ lib/i18n/lt.js
@@ -1,0 +1,13 @@
+/*jslint node: true */
+"use strict";
+
+module.exports = {
+    ARTISTS_FOLDER : "Atlikėjai",
+
+    UNKNOWN_ARTIST : "Nežinomas atlikėjas",
+    UNKNOWN_ALBUM : "Nežinomas albumas",
+    UNKNOWN_GENRE : "Nežinomas žanras",
+    UNKNOWN_TITLE : "Nežinomas pavadinimas",
+
+    ROOT_NAME: "Tėvinis"
+};
