+++ lib/upnpServer.js
@@ -31,10 +31,10 @@ var UpnpServer = function(port, _configuration, callback) {
 
   var langPart = /^([a-z][a-z]).*$/i.exec(lang);
   try {
-    configuration.i18n = require("./i18n_" + langPart[1].toLowerCase());
+    configuration.i18n = require("./i18n/" + langPart[1].toLowerCase());
   } catch(e) {
     // if localization is not supported, trying to use english by default
-    configuration.i18n = require("./i18n_en");
+    configuration.i18n = require("./i18n/en");
   }
 
   this.dlnaSupport = configuration.dlna;
